Refs #38: Fetch Data From Firestore

+++ src/pages/Menu/MenuPage.jsx
@@ -1,106 +1,89 @@
 import { Fab, Grid, Typography, Button } from "@mui/material";
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import ItemCard from "../../components/ItemCard/ItemCard";
 import AddSectionModal from "../../components/Modals/AddSectionModal";
 import ResponsiveDrawer from "../../components/Sidebar/Sidebar";
 import CreateMenuPage from "../RestaurantCreation/CreateMenu/CreateMenuPage";
+import SearchOffIcon from '@mui/icons-material/SearchOff';
 import { SectionStyled } from "./style";
+import { collection, doc, getDocs, query, updateDoc, where } from "firebase/firestore";
+import { db } from "../../../firebase.config";
+import { grey } from "@mui/material/colors";
 
 export default function MenuPage() {
-  const [currentSection, setCurrentSection] = useState("");
-  // REPLACE THIS WITH ACTUAL DATA FROM FIRESTORE
-  const [items, setItems] = useState([
-    {
-      availability: true,
-      name: "Noodle Soup",
-      price: 15.5,
-      options: [
-        {
-          availability: true,
-          name: "Rare Beef",
-          price: 0,
-        },
-        {
-          availability: true,
-          name: "Brisket",
-          price: 0,
-        },
-        {
-          availability: true,
-          name: "Meat Ball",
-          price: 0,
-        },
-      ],
-    },
-    {
-      availability: true,
-      name: "Wonton Soup",
-      price: 17.5,
-      options: [
-        {
-          availability: true,
-          name: "ADD-Wonton",
-          price: 3,
-        },
-      ],
-    },
-    {
-      availability: true,
-      name: "Bun Bo Hue",
-      price: 15.5,
-      options: [
-        {
-          availability: true,
-          name: "ADD-Rare Beef",
-          price: 4,
-        },
-        {
-          availability: true,
-          name: "ADD-Brisket",
-          price: 4,
-        },
-        {
-          availability: true,
-          name: "ADD-Meat Ball",
-          price: 4,
-        },
-      ],
-    },
-  ]);
+  const [docId, setDocId] = useState("");
+  const [menuData, setMenuData] = useState([]);
+  const [currentSection, setCurrentSection] = useState({});
   const [newSectionName, setNewSectionName] = useState("");
   const [openAddSectionModal, setOpenAddSectionModal] = useState(false);
   const [openCreateMenuForm, setOpenCreateMenuForm] = useState(false);
-  // REPLACE THIS WITH ACTUAL DATA FROM FIRESTORE
-  const [sections, setSections] = useState([
-    "Bar",
-    "Grill",
-    "Apetizer",
-    "Soup",
-  ]);
+
+  // Get Menu Doc Ref
+  const menuCollection = collection(db, "menu");
+  const restaurantRef = JSON.parse(
+    localStorage.getItem("current-user")
+  ).docId;
+  const menuRef = query(
+    menuCollection,
+    where("restaurantRef", "==", `/users/${restaurantRef}`)
+  );
 
   const addSection = (section) => {
-    const newSections = [...sections];
-    newSections.push(section);
-    setSections(newSections);
+    const menu = [...menuData];
+    menu.push({name: section});
+    setMenuData(menu);
   };
-
+  
   const deleteItem = (index) => {
-    const newItems = [...items];
-    newItems.splice(index, 1);
-    setItems(newItems);
+    const menu = [...menuData];
+    menu[currentSection.index].items.splice(index, 1);
+    setMenuData(menu);
   };
-
+  
   const handleChangeItems = (targetItem, field, value) => {
+    const menu = [...menuData];
+    let items = menu[currentSection.index].items;
     const newItems = items.map((item) => {
-      if (item.name === targetItem.name) {
+      if (item.itemName === targetItem.itemName) {
         return { ...item, [field]: value };
       } else {
         return item;
       }
     });
-    setItems(newItems);
+    menu[currentSection.index].items = newItems;
+    setMenuData(menu)
   };
+  
+  useEffect(() => {
+    const fetchMenu = async () => {
+      try {
+        const querySnapshot = await getDocs(menuRef);
+        const newMenu = [];
+        querySnapshot.forEach((doc) => {
+          setDocId(doc.id);
+          newMenu.push(...doc.data().sections);
+        });
+        setMenuData(newMenu);
+        setCurrentSection({name: newMenu[0].name, index: 0})
+      } catch (error) {
+        console.log(error);
+      }
+    };
+    fetchMenu();
+  }, []);
 
+  useEffect(() => {
+    const handleChangeDocument = async () => {
+      try {
+        const docRef = doc(db, "menu", docId);
+        await updateDoc(docRef, {sections: menuData});
+      } catch(error) {
+        console.log(error);
+      }
+    }
+    handleChangeDocument();
+  }, [menuData])
+  
   const menu = (
     <>
       {openCreateMenuForm ? (
@@ -133,16 +116,18 @@ export default function MenuPage() {
             >
               <Grid item xs={12} md={9}>
                 <Grid container columnSpacing={5}>
-                  {sections.map((section, index) => {
+                  {menuData.map((section, index) => {
                     return (
                       <Grid item key={index}>
                         <SectionStyled
-                          currentSection={currentSection == section}
+                          currentSection={currentSection.name == section.name}
                           padding={1}
-                          onClick={() => setCurrentSection(section)}
+                          onClick={() =>
+                            setCurrentSection({ name: section.name, index })
+                          }
                           variant="h5"
                         >
-                          {section}
+                          {section.name}
                         </SectionStyled>
                       </Grid>
                     );
@@ -178,16 +163,38 @@ export default function MenuPage() {
               padding={2}
               rowGap={2}
             >
-              {items.map((item, index) => {
-                return (
-                  <Grid key={index} item xs={12} md={5} xl={3}>
-                    <ItemCard
-                      deleteItem={() => deleteItem(index)}
-                      item={item}
-                      setItem={handleChangeItems}
-                    />
-                  </Grid>
-                );
+              {menuData.map((section, index) => {
+                if (currentSection.name === section.name) {
+                  if (!section.items) {
+                    return (
+                      <Grid
+                        alignItems="center"
+                        container
+                        justifyContent="center"
+                        rowGap={2}
+                      >
+                        <Grid item textAlign="center" xs={12}>
+                          <SearchOffIcon fontSize="large" sx={{color: grey[500]}} />
+                        </Grid>
+                        <Grid item textAlign="center" xs={12}>
+                          <Typography fontWeight="bold" variant="h4" sx={{color: grey[500]}}>
+                            THIS SECTION IS EMPTY. LETS ADD SOME ITEMS INTO IT
+                          </Typography>
+                        </Grid>
+                      </Grid>
+                    );
+                  } else {
+                    return section.items.map((item, index) => (
+                      <Grid key={index} item xs={12} md={5} xl={3}>
+                        <ItemCard
+                          deleteItem={() => deleteItem(index)}
+                          item={item}
+                          setItem={handleChangeItems}
+                        />
+                      </Grid>
+                    ));
+                  }
+                }
               })}
             </Grid>
             <Grid container justifyContent="flex-end" padding={2}>
